Create a Hijri module and move hijri_ummalqura_data to it
