+++ components/calendar/src/cal/hijri.rs
@@ -19,7 +19,7 @@
 //! assert_eq!(hijri_date.day_of_month().0, 11);
 //! ```
 
-use crate::cal::hijri_ummalqura_data::{UMMALQURA_DATA, UMMALQURA_DATA_STARTING_YEAR};
+use crate::cal::hijri::ummalqura_data::{UMMALQURA_DATA, UMMALQURA_DATA_STARTING_YEAR};
 use crate::cal::iso::{Iso, IsoDateInner};
 use crate::calendar_arithmetic::PrecomputedDataSource;
 use crate::calendar_arithmetic::{ArithmeticDate, CalendarArithmetic};
@@ -34,6 +34,8 @@ use calendrical_calculations::rata_die::RataDie;
 use icu_provider::marker::ErasedMarker;
 use icu_provider::prelude::*;
 use tinystr::tinystr;
+
+mod ummalqura_data;
 
 fn era_year(year: i32) -> EraYear {
     types::EraYear {
